+++ addon/components/fd-sheets/fd-base-sheet.js
@@ -1,0 +1,133 @@
+import Component from '@ember/component';
+import FdSaveHasManyRelationshipsMixin from '../../mixins/fd-save-has-many-relationships';
+import { inject as service } from '@ember/service';
+import { assert } from '@ember/debug';
+
+export default Component.extend(FdSaveHasManyRelationshipsMixin, {
+
+  /**
+   Service that get current project contexts.
+
+   @property currentProjectContext
+   @type {Class}
+   @default service()
+   */
+  currentProjectContext: service('fd-current-project-context'),
+
+  /**
+    Service for managing the state of the application.
+
+    @property appState
+    @type AppStateService
+  */
+  appState: service(),
+
+  /**
+    @property store
+    @type Service
+  */
+  store: service(),
+
+  /**
+    Service for managing the state of the component.
+
+    @property fdSheetService
+    @type FdSheetService
+  */
+  fdSheetService: service(),
+
+  /**
+    Service for managing the state of the component.
+
+    @property fdDialogService
+    @type fdDialogService
+  */
+  fdDialogService: service('fd-dialog-service'),
+
+  /**
+    Sheet component name.
+
+    @property sheetComponentName
+    @type String
+    @default ''
+  */
+  sheetComponentName: '',
+
+  /**
+    Flag: indicates whether property is readonly.
+
+    @property readonlyMode
+    @type Boolean
+    @default true
+   */
+  readonlyMode: true,
+
+  init() {
+    this._super(...arguments);
+
+    this.set('classNames', ['fd-sheet', `${this.get('sheetComponentName')}`]);
+    this.set('classNameBindings',
+    [
+      `fdSheetService.sheetSettings.visibility.${this.get('sheetComponentName')}:visible`,
+      `fdSheetService.sheetSettings.expanded.${this.get('sheetComponentName')}:expand`,
+    ]);
+
+    this.get('fdSheetService').on('openSheetTriggered', this, this.openSheetBase);
+    this.get('fdSheetService').on('closeSheetTriggered', this, this.closeSheetBase);
+  },
+
+  willDestroy() {
+    this._super(...arguments);
+
+    this.get('fdSheetService').off('openSheetTriggered', this, this.openSheetBase);
+    this.get('fdSheetService').off('closeSheetTriggered', this, this.closeSheetBase);
+  },
+
+  /**
+    Opening sheet base method.
+
+     @method openSheetBase
+     @param {String} sheetName Sheet's dbName
+     @param {Object} currentItem Current list item
+  */
+  openSheetBase(sheetName, currentItem) {
+    let sheetComponentName = this.get('sheetComponentName');
+    if (sheetComponentName === sheetName) {
+      this.openSheet(sheetName, currentItem);
+    }
+  },
+
+  /**
+    Closing sheet base method.
+
+     @method closeSheetBase
+     @param {String} sheetName Sheet's dbName
+  */
+  closeSheetBase(sheetName) {
+    let sheetComponentName = this.get('sheetComponentName');
+    if (sheetComponentName === sheetName) {
+      this.closeSheet(sheetName);
+    }
+  },
+
+  /**
+    Opening sheet.
+
+     @method openSheet
+     @param {String} sheetName Sheet's dbName
+     @param {Object} currentItem Current list item
+  */
+  openSheet(sheetName) {
+    assert(`Please specify 'openSheet' method for '${sheetName}' sheet compoenent`);
+  },
+
+  /**
+    Closing sheet.
+
+     @method closeSheet
+     @param {String} sheetName Sheet's dbName
+  */
+  closeSheet(sheetName) {
+    assert(`Please specify 'closeSheet' method for '${sheetName}' sheet compoenent`);
+  },
+});
