+++ addon/components/fd-sheets/fd-sheets-tool-bar.js
@@ -1,0 +1,149 @@
+import Component from '@ember/component';
+import { inject as service } from '@ember/service';
+import { computed } from '@ember/object';
+import layout from '../../templates/components/fd-sheets/fd-sheets-tool-bar';
+
+export default Component.extend({
+  layout,
+
+  /**
+    Sheet component name.
+
+    @property sheetComponentName
+    @type String
+    @default ''
+  */
+  sheetComponentName: '',
+
+  /**
+    Service for managing the state of the component.
+
+    @property fdSheetService
+    @type FdSheetService
+  */
+  fdSheetService: service(),
+
+  /**
+    Array button.
+
+    @property sheetButtons
+    @type Ember.A()
+  */
+  sheetButtons: undefined,
+
+  /**
+    Object visible button.
+
+    @property buttonArray
+    @type Ember.A()
+  */
+  buttonVisible: computed('sheetButtons', function() {
+    let sheetButtons = this.get('sheetButtons');
+    let buttons = {};
+    sheetButtons.forEach((button) => {
+      buttons[`${button}`] = true;
+    });
+
+    return buttons;
+  }),
+
+  /**
+    Custom button visible.
+
+    @property customButton
+    @type Bool
+  */
+  customButtonVisible: false,
+
+  /**
+    Custom button title.
+
+    @property customButton
+    @type String
+  */
+  customButtonTitle: undefined,
+
+  /**
+    Flag: indicates whether to show toolbar.
+
+    @property toolbarVisible
+    @type Bool
+  */
+  toolbarVisible: true,
+
+  /**
+    Flag: indicates whether property is readonly.
+
+    @property readonlyMode
+    @type Boolean
+    @default false
+   */
+  readonlyMode: false,
+
+  /**
+    Flag: indicates whether to not show button for new model.
+
+    @property isNewModel
+    @type Bool
+  */
+  isNewModel: false,
+
+  actions: {
+    /**
+      Closing sheet.
+
+      @method actions.closeSheet
+    */
+    closeSheet() {
+      let sheetComponentName = this.get('sheetComponentName');
+      this.get('fdSheetService').closeSheet(sheetComponentName);
+    },
+
+    /**
+      Expanding sheet.
+
+      @method actions.expand
+    */
+    expand() {
+      let sheetComponentName = this.get('sheetComponentName');
+      this.get('fdSheetService').expand(sheetComponentName);
+    },
+
+    /**
+      Saving changes.
+
+      @method actions.save
+    */
+    save() {
+      this.get('saveSheet')();
+      this.set('readonlyMode', true);
+    },
+
+    /**
+      Delete data.
+
+      @method actions.delete
+    */
+    delete() {
+      this.get('deleteSheet')();
+    },
+
+    /**
+      Off readonlyMode.
+
+      @method actions.edit
+    */
+    edit() {
+      this.set('readonlyMode', false);
+    },
+
+    /**
+      Custom button action.
+
+      @method actions.customButtonAction
+    */
+    customButtonAction() {
+      this.get('customButtonSheet')();
+    },
+  }
+});
