+++ addon/controllers/fd-application-model.js
@@ -1,18 +1,12 @@
 import Controller from '@ember/controller';
-import EmberObject, { computed, observer, set } from '@ember/object';
+import EmberObject, { computed, observer } from '@ember/object';
 import { isBlank, isNone } from '@ember/utils';
 import { A } from '@ember/array';
 import { inject as service } from '@ember/service';
-import { updateStrByObjects } from '../utils/fd-update-str-value';
-import { translationMacro as t } from 'ember-i18n';
-import FdSheetCloseConfirm from '../mixins/fd-sheet-close-confirm';
 import $ from 'jquery';
-import { resolve, reject } from 'rsvp';
-import { createClassPrimitive, deletePrimitives } from '../utils/fd-update-class-diagram';
+import FdSheetCloseConfirm from '../mixins/fd-sheet-close-confirm';
 
-import FdSaveHasManyRelationshipsMixin from '../mixins/fd-save-has-many-relationships';
-
-export default Controller.extend(FdSaveHasManyRelationshipsMixin, FdSheetCloseConfirm, {
+export default Controller.extend(FdSheetCloseConfirm, {
 
   /**
    Service that get current project contexts.
@@ -40,22 +34,6 @@ export default Controller.extend(FdSaveHasManyRelationshipsMixin, FdSheetCloseCo
   fdSheetService: service(),
 
   /**
-    Value selected entity.
-
-    @property selectedElement
-    @type Object
-  */
-  selectedElement: undefined,
-
-  /**
-    Value selected view.
-
-    @property selectedView
-    @type Object
-  */
-  selectedView: undefined,
-
-  /**
     Value search input.
 
     @property searchValue
@@ -72,23 +50,6 @@ export default Controller.extend(FdSaveHasManyRelationshipsMixin, FdSheetCloseCo
     @default ''
   */
   sheetComponentName: '',
-
-  /**
-    Sheet view name.
-
-    @property sheetViewName
-    @type String
-    @default ''
-  */
-  sheetViewName: '',
-
-  /**
-    Object with flags indicates whether edit panel is readonly.
-
-    @property readonlyMode
-    @type Object
-  */
-  readonlyMode: undefined,
 
   /**
     Flag: indicates whether to show create editing panel.
@@ -112,41 +73,6 @@ export default Controller.extend(FdSaveHasManyRelationshipsMixin, FdSheetCloseCo
     }
 
     this.closeViewSheet();
-  }),
-
-  /**
-    Sheet title value.
-
-    @method computedTitle
-  */
-  computedTitle: computed('isAddMode', 'i18n.locale', 'selectedElement', {
-    get() {
-      return this.get('isAddMode') ? t('components.fd-create-entity.caption') : this.get('selectedElement.model.data.name');
-    },
-    set(key, value) {
-      if (!this.get('isAddMode')) {
-        this.set('selectedElement.model.data.name', value);
-      }
-
-      return value;
-    }
-  }),
-
-  /**
-    Removes quotes from class stereotype.
-
-    @method componentNamePart
-  */
-  componentNamePart: computed('selectedElement', function() {
-    let selectedElement = this.get('selectedElement');
-    if (!isNone(selectedElement)) {
-      let stereotype = selectedElement.get('model.data.stereotype');
-      if (isBlank(stereotype)) {
-        return 'implementation';
-      }
-
-      return stereotype.substring(1, stereotype.length - 1);
-    }
   }),
 
   /**
@@ -222,300 +148,7 @@ export default Controller.extend(FdSaveHasManyRelationshipsMixin, FdSheetCloseCo
     return newModel;
   }),
 
-  init() {
-    this._super(...arguments);
-
-    this.get('fdSheetService').on('openSheetTriggered', this, this.openSheet);
-    this.get('fdSheetService').on('closeSheetTriggered', this, this.closeSheet);
-  },
-
-  /**
-    Deactivate item from selectedElement.
-
-     @method deactivateListItem
-  */
-  deactivateListItem() {
-    let selectedElement = this.get('selectedElement');
-    if (!isNone(selectedElement)) {
-      let model = selectedElement.get('model.data');
-      model.rollbackAll();
-      selectedElement.set('model.active', false);
-    }
-  },
-
-  /**
-    Update bs value in model.
-
-     @method updateClassModel
-  */
-  updateClassModel(modelSelectedElement) {
-    let stereotype = modelSelectedElement.get('stereotype');
-    if (stereotype === '«implementation»' || isBlank(stereotype)) {
-      let model = this.get('model');
-      let classObj = model.classes.findBy('settings.data.id', modelSelectedElement.id);
-      let bs = modelSelectedElement.get('businessServerClass');
-      let bsModel = isNone(bs) ? null : { data: bs, active: false };
-      set(classObj, 'bs', bsModel);
-    }
-  },
-
-  /**
-    Opening sheet.
-
-     @method openSheet
-     @param {String} sheetName Sheet's dbName
-     @param {Object} currentItem Current list item
-  */
-  openSheet(sheetName, currentItem) {
-    let sheetComponentName = this.get('sheetComponentName');
-    if (sheetComponentName === sheetName) {
-      this.deactivateListItem();
-      this.closeViewSheet();
-      this.set('readonlyMode.sheetComponentName', true);
-      this.set('selectedElement', currentItem);
-      if (!isNone(currentItem)) {
-        this.set('isAddMode', false);
-      }
-    }
-  },
-
-  /**
-    Closing sheet.
-
-     @method closeSheet
-     @param {String} sheetName Sheet's dbName
-  */
-  closeSheet(sheetName) {
-    let sheetComponentName = this.get('sheetComponentName');
-    let sheetViewName = this.get('sheetViewName');
-    if (sheetComponentName === sheetName) {
-      this.deactivateListItem();
-      this.closeViewSheet();
-      this.set('readonlyMode.sheetComponentName', true);
-      this.set('selectedElement', undefined);
-    } else if (sheetViewName === sheetName) {
-      let selectedView = this.get('selectedView');
-      if (!isNone(selectedView)) {
-        selectedView.rollbackAll();
-        this.set('readonlyMode.sheetViewName', true);
-        this.set('selectedView', undefined);
-      }
-    }
-  },
-
-  /**
-    Closing view sheet.
-
-     @method closeViewSheet
-  */
-  closeViewSheet() {
-    let sheetViewName = this.get('sheetViewName');
-    let fdSheetService = this.get('fdSheetService');
-    if (fdSheetService.isVisible(sheetViewName)) {
-      fdSheetService.closeSheet(sheetViewName);
-    }
-  },
-
-  /**
-    Find array in model.
-
-     @method getModelArrayByStereotype
-     @param {Object} model model object
-  */
-  getModelArrayByStereotype(model) {
-    let findArray;
-    switch (model.get('stereotype')) {
-      case null:
-      case '':
-      case '«implementation»':
-        findArray = this.get('model.classes');
-        break;
-      case '«listform»': {
-        let dataObjectId = model.get('formViews.firstObject.view.class.id');
-        let classArray = this.get('model.classes').findBy('settings.data.id', dataObjectId);
-        findArray = classArray.listForms;
-        break;
-      }
-      case '«editform»': {
-        let dataObjectId = model.get('formViews.firstObject.view.class.id');
-        let classArray = this.get('model.classes').findBy('settings.data.id', dataObjectId);
-        findArray = classArray.editForms;
-        break;
-      }
-      case '«enumeration»':
-        findArray = this.get('model.enums');
-        break;
-      case '«typedef»':
-        findArray = this.get('model.typedefs');
-        break;
-      case '«type»':
-        findArray = this.get('model.types');
-        break;
-      case '«application»':
-        findArray = this.get('model.applications');
-        break;
-      case '«businessserver»':
-        findArray = this.get('model.bs');
-        break;
-      case '«external»':
-        findArray = this.get('model.externals');
-        break;
-      case '«externalinterface»':
-        findArray = this.get('model.externalinterface');
-        break;
-      case '«interface»':
-        findArray = this.get('model.interfaces');
-        break;
-      case '«userform»':
-        findArray = this.get('model.userforms');
-        break;
-      case '«geolayer»':
-        findArray = this.get('model.geolayers');
-        break;
-      case '«geolayerstyle»':
-        findArray = this.get('model.geolayerstyles');
-        break;
-      default:
-        findArray = this.get('model.userstereotypes');
-    }
-
-    return findArray;
-  },
-
-  /**
-    Add in model new classes.
-
-     @method addNewClassInModel
-  */
-  addNewClassInModel() {
-    let model = this.get('selectedElement.model');
-    let modelHash = this.getModelArrayByStereotype(model.data);
-
-    if (model.data.get('stereotype') === '«implementation»') {
-      modelHash.pushObject({ settings: model, editForms: A(), listForms: A(), parents: A(), bs: null });
-    } else {
-      modelHash.pushObject(model);
-    }
-
-    this.notifyPropertyChange('model');
-  },
-
-  /**
-    Check data for correctness.
-
-    @method validateData
-    @param {Object} model class model.
-  */
-  validateData(model) {
-    let modelName = model.get('name');
-
-    if (isBlank(modelName)) {
-      return reject({ message: this.get('i18n').t('forms.fd-application-model.error-message.empty-class').toString() });
-    }
-
-    if ((model.get('stereotype') === '«editform»' || model.get('stereotype') === '«listform»') && isNone(model.get('formViews.firstObject.view'))) {
-      return reject({ message: this.get('i18n').t('forms.fd-application-model.error-message.view-form').toString() });
-    }
-
-    if (model.get('isNew')) {
-      model.set('nameStr', model.get('name'));
-      // Get current classes.
-      let allClasses = this.get('store').peekAll('fd-dev-class');
-      let classesCurrentStage = allClasses.filterBy('stage.id', this.get('currentProjectContext').getCurrentStage());
-      let currentClass = A(classesCurrentStage).find((a) => {
-        return a.get('name') === modelName && !isNone(a.get('id'));
-      });
-
-      if (!isNone(currentClass)) {
-        return reject({ message: this.get('i18n').t('forms.fd-application-model.error-message.exist-class').toString() });
-      }
-    }
-
-    return resolve();
-  },
-
   actions: {
-
-    /**
-      Save 'selectedElement'.
-
-       @method actions.save
-       @param {Boolean} closeAfter close after save
-    */
-    save(closeAfter) {
-      const selectedElement = this.get('selectedElement');
-      let model = selectedElement.get('model.data');
-      this.get('appState').loading();
-      updateStrByObjects(model);
-
-      let isNew = model.get('isNew');
-      this.validateData(model)
-      .then(() => model.save())
-      .then(() => this.saveHasManyRelationships(model))
-      .then(() => {
-        if (isNew) {
-          this.addNewClassInModel();
-          let diagram = createClassPrimitive(this.get('store'), this.get('currentProjectContext'), model);
-          return diagram.save();
-        }
-
-        return resolve();
-      })
-      .then(() => {
-        this.updateClassModel(model);
-        if (closeAfter) {
-          this.confirmClose(this.get('sheetComponentName'));
-        }
-      })
-      .catch((error) => {
-        this.set('isError', true);
-        this.set('messageText', error.message);
-        this.set('show', true);
-      })
-      .finally(() => {
-        this.get('appState').reset();
-      });
-    },
-
-    /**
-      Save 'selectedView'.
-
-       @method actions.saveView
-    */
-    saveView() {
-      let view = this.get('selectedView');
-      this.get('appState').loading();
-      view.save()
-      .catch((error) => {
-        this.set('isError', true);
-        this.set('messageText', error.message);
-        this.set('show', true);
-      })
-      .finally(() => {
-        this.get('appState').reset();
-      });
-    },
-
-    /**
-      Opening sheet 'view-sheet'.
-
-       @method actions.openViewSheet
-       @param {Object} view view
-    */
-    openViewSheet(view) {
-      this.set('selectedView', view);
-      this.get('fdSheetService').openSheet(this.get('sheetViewName'));
-    },
-
-    /**
-      Open create class edit panel.
-
-       @method actions.openCreateClassEditPanel
-    */
-    openCreateClassEditPanel() {
-      this.set('isAddMode', true);
-      this.get('fdSheetService').openSheet(this.get('sheetComponentName'));
-    },
 
     /**
       Create new class.
@@ -525,8 +158,6 @@ export default Controller.extend(FdSaveHasManyRelationshipsMixin, FdSheetCloseCo
        @param {Object} dataobject dataobject for list and edit form.
     */
     createClass(stereotype, dataobject) {
-      this.deactivateListItem();
-      this.set('selectedElement', undefined);
       let store = this.get('store');
       let currentStage = this.get('currentProjectContext').getCurrentStageModel();
       let newClass = store.createRecord('fd-dev-class', {
@@ -554,89 +185,22 @@ export default Controller.extend(FdSaveHasManyRelationshipsMixin, FdSheetCloseCo
     },
 
     /**
-      Delete selected class.
+      Open create class edit panel.
 
-       @method actions.delete
+       @method actions.openCreateClassEditPanel
     */
-    delete() {
-      const store = this.get('store');
-      let selectedElement = this.get('selectedElement.model.data');
-      let modelHash = this.getModelArrayByStereotype(selectedElement);
-
-      let deleteObject;
-      let deleteModels = A();
-      let stereotype = selectedElement.get('stereotype');
-      if (stereotype === '«businessserver»') {
-        let bsInClass = this.get('model.classes').filterBy('bs.data.id', selectedElement.id);
-        if (bsInClass.length > 0) {
-          this.set('isError', true);
-          this.set('messageText', this.get('i18n').t('forms.fd-application-model.error-message.exist-class').toString() + A(bsInClass).get('firstObject.settings.data.name'));
-          this.set('show', true);
-          return;
-        }
-      }
-
-      if (stereotype === '«implementation»' || isBlank(stereotype)) {
-        let deleteObjectClass = modelHash.findBy('settings.data.id', selectedElement.id);
-        deleteModels.pushObjects(deleteObjectClass.listForms);
-        deleteModels.pushObjects(deleteObjectClass.editForms);
-        deleteObject = deleteObjectClass.settings;
-        modelHash.removeObject(deleteObjectClass);
-      } else {
-        deleteObject = modelHash.findBy('data.id', selectedElement.id);
-        modelHash.removeObject(deleteObject);
-      }
-
-      this.get('appState').loading();
-
-      deleteModels.pushObject(deleteObject);
-      let modelsForBatchUpdate = deleteModels.map((a) => {
-        let data = a.data;
-        data.deleteRecord();
-
-        return data;
-      });
-
-      let primitivesOnDelete = deletePrimitives(store, this.get('currentProjectContext'), deleteModels.map((a) => a.data));
-      A(modelsForBatchUpdate).pushObjects(primitivesOnDelete);
-
-      store.batchUpdate(modelsForBatchUpdate)
-      .then(() => {
-        this.set('selectedElement', undefined);
-        this.get('fdSheetService').closeSheet(this.get('sheetComponentName'));
-      })
-      .catch((error) => {
-        this.set('isError', true);
-        this.set('messageText', error.message);
-        this.set('show', true);
-      })
-      .finally(() => {
-        this.get('appState').reset();
-      });
+    openCreateClassEditPanel() {
+      this.set('isAddMode', true);
+      this.get('fdSheetService').openSheet(this.get('sheetComponentName'));
     },
 
     /**
-      Delete selected view.
+      This method will notify all observers that the model changed value
 
-       @method actions.deleteView
+       @method actions.updateModel
     */
-    deleteView() {
-      let view = this.get('selectedView');
-
-      this.get('appState').loading();
-      view.destroyRecord()
-      .then(() => {
-        this.set('selectedView', undefined);
-        this.get('fdSheetService').closeSheet(this.get('sheetViewName'));
-      })
-      .catch((error) => {
-        this.set('isError', true);
-        this.set('messageText', error.message);
-        this.set('show', true);
-      })
-      .finally(() => {
-        this.get('appState').reset();
-      });
+    updateModel() {
+      this.notifyPropertyChange('model');
     }
   }
 });
