+++ addon/locales/en/components/fd-sheets-tool-bar.js
@@ -1,0 +1,5 @@
+export default {
+  'placeholder-caption': 'Enter name',
+  'save-caption': 'Save',
+  'edit-caption': 'Edit'
+};
