+++ addon/locales/ru/components/fd-sheets-tool-bar.js
@@ -1,0 +1,5 @@
+export default {
+  'placeholder-caption': 'Введите имя',
+  'save-caption': 'Сохранить',
+  'edit-caption': 'Правка'
+};
