+++ addon/locales/ru/translations.js
@@ -138,6 +138,7 @@ import FdCreateView from './components/fd-create-view';
 import FdCreateEntity from './components/fd-create-entity';
 import FdCreateDiagrams from './components/fd-create-diagrams';
 import FdSheet from './components/fd-sheet';
+import FdSheetToolBar from './components/fd-sheets-tool-bar';
 import FdCreateNodeEditingPanel from './components/fd-create-node-editing-panel';
 import FdFolderEditingPanel from './components/fd-folder-editing-panel';
 import FdPropertyEditingPanel from './components/fd-property-editing-panel';
@@ -450,6 +451,7 @@ $.extend(true, translations, {
     'fd-geolayerstyle-editing-panel': FdGeolayerstyleFormEditingPanel,
     'fd-create-view': FdCreateView,
     'fd-sheet': FdSheet,
+    'fd-sheets-tool-bar': FdSheetToolBar,
     'fd-create-node-editing-panel': FdCreateNodeEditingPanel,
     'fd-folder-editing-panel': FdFolderEditingPanel,
     'fd-property-editing-panel': FdPropertyEditingPanel,
