+++ app/components/fd-sheets/fd-base-sheet.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-designer/components/fd-sheets/fd-base-sheet';
