+++ app/components/fd-sheets/fd-sheets-tool-bar.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-designer/components/fd-sheets/fd-sheets-tool-bar';
