+++ app/services/fd-dialog-service.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-designer/services/fd-dialog-service';
